+++ l10n_se_skv_vat_report/__init__.py
@@ -1,4 +1,5 @@
 from . import (
+    models,  # noqa: F401
     reports,  # noqa: F401
     wizard,  # noqa: F401
 )